+++ packages/shared/src/UI/components/SettingsBoard/GasForm.tsx
@@ -7,76 +7,11 @@ import { Grid, Typography } from '@mui/material'
 import { WarningIcon } from '@masknet/icons'
 import { useSharedI18N } from '@masknet/shared'
 import type { ChainId, GasOption, Transaction } from '@masknet/web3-shared-evm'
-import {
-    GasOptionType,
-    isGreaterThanOrEqualTo,
-    isLessThan,
-    isLessThanOrEqualTo,
-    isPositive,
-    multipliedBy,
-    NetworkPluginID,
-} from '@masknet/web3-shared-base'
+import { GasOptionType, NetworkPluginID } from '@masknet/web3-shared-base'
 import { useWeb3State } from '@masknet/plugin-infra/web3'
-import { z as zod } from 'zod'
+import type { z as zod } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
-
-const HIGH_FEE_WARNING_MULTIPLIER = 1.5
-
-function useSchema(chainId: ChainId, transaction: Transaction, gasOptions: Record<GasOptionType, GasOption>) {
-    const t = useSharedI18N()
-    const { Others } = useWeb3State(NetworkPluginID.PLUGIN_EVM)
-    const isEIP1559 = Others?.chainResolver.isSupport(chainId, 'EIP1559')
-    return useMemo(() => {
-        return zod
-            .object({
-                gasLimit: zod
-                    .string()
-                    .min(1, t.gas_settings_error_gas_limit_absence())
-                    .refine(
-                        (gasLimit) => isGreaterThanOrEqualTo(gasLimit, transaction.gas as string),
-                        t.gas_settings_error_min_gas_limit_tips({ limit: transaction.gas as string }),
-                    ),
-                gasPrice: zod.string().min(isEIP1559 ? 0 : 1, t.gas_settings_error_gas_price_absence()),
-                maxPriorityFeePerGas: zod
-                    .string()
-                    .min(1, t.gas_settings_error_max_priority_fee_absence())
-                    .refine(isPositive, t.gas_settings_error_max_priority_gas_fee_positive())
-                    .refine((value) => {
-                        return isGreaterThanOrEqualTo(value, gasOptions.slow?.suggestedMaxPriorityFeePerGas ?? 0)
-                    }, t.gas_settings_error_max_priority_gas_fee_too_low())
-                    .refine(
-                        (value) =>
-                            isLessThan(
-                                value,
-                                multipliedBy(
-                                    gasOptions.fast?.suggestedMaxPriorityFeePerGas ?? 0,
-                                    HIGH_FEE_WARNING_MULTIPLIER,
-                                ),
-                            ),
-                        t.gas_settings_error_max_priority_gas_fee_too_high(),
-                    ),
-                maxFeePerGas: zod
-                    .string()
-                    .min(1, t.gas_settings_error_max_fee_absence())
-                    .refine(
-                        (value) => isGreaterThanOrEqualTo(value, gasOptions.slow?.suggestedMaxFeePerGas ?? 0),
-                        t.gas_settings_error_max_fee_too_low(),
-                    )
-                    .refine(
-                        (value) =>
-                            isLessThan(
-                                value,
-                                multipliedBy(gasOptions.fast?.suggestedMaxFeePerGas ?? 0, HIGH_FEE_WARNING_MULTIPLIER),
-                            ),
-                        t.gas_settings_error_max_fee_too_high(),
-                    ),
-            })
-            .refine((data) => isLessThanOrEqualTo(data.maxPriorityFeePerGas, data.maxFeePerGas), {
-                message: t.gas_settings_error_max_priority_gas_fee_imbalance(),
-                path: ['maxFeePerGas'],
-            })
-    }, [t, isEIP1559, transaction.gas, gasOptions, Others])
-}
+import { useGasSchema } from './hooks'
 
 const useStyles = makeStyles()((theme) => {
     return {
@@ -106,7 +41,7 @@ export interface GasFormProps {
     chainId: ChainId
     transaction: Transaction
     gasOptions: Record<GasOptionType, GasOption>
-    onChange?: (data?: Partial<zod.infer<ReturnType<typeof useSchema>>>) => void
+    onChange?: (data?: Partial<zod.infer<ReturnType<typeof useGasSchema>>>) => void
 }
 
 export function GasForm(props: GasFormProps) {
@@ -117,7 +52,7 @@ export function GasForm(props: GasFormProps) {
 
     const isEIP1559 = Others?.chainResolver.isSupport(chainId, 'EIP1559')
 
-    const schema = useSchema(chainId, transaction, gasOptions)
+    const schema = useGasSchema(chainId, transaction, gasOptions)
 
     const methods = useForm<zod.infer<typeof schema>>({
         shouldUnregister: false,
